refactor: move environment code into client dir
