+++ src/client/features/location.test.ts
@@ -1,5 +1,5 @@
 import * as assert from 'assert'
-import { TextDocumentLocationProviderRegistry } from '../../environment/providers/location'
+import { TextDocumentLocationProviderRegistry } from '../providers/location'
 import { ClientCapabilities, DefinitionRequest, ReferenceParams, TextDocumentPositionParams } from '../../protocol'
 import { Client } from '../client'
 import {
