+++ src/client/providers/hover.ts
