+++ src/client/providers/registry.test.ts
