+++ src/client/providers/registry.ts
