+++ src/client/registries.test.ts
