+++ src/client/registries.ts
