+++ src/extension/extensionHost.integration.test.ts
@@ -3,9 +3,9 @@ import { from } from 'rxjs'
 import { filter, map, switchMap, take } from 'rxjs/operators'
 import { MarkupKind } from 'sourcegraph'
 import * as sourcegraph from 'sourcegraph'
+import { Controller } from '../client/controller'
+import { Environment } from '../client/environment'
 import { HoverMerged } from '../client/types/hover'
-import { Controller } from '../environment/controller'
-import { Environment } from '../environment/environment'
 import { clientStateIsActive } from '../test/helpers'
 import { createMessageTransports } from '../test/integration/helpers'
 import { createExtensionHost } from './extensionHost'
