+++ src/extension/old/api.ts
@@ -1,7 +1,7 @@
 import { Subscription } from 'rxjs'
 import * as sourcegraph from 'sourcegraph'
+import { Context } from '../../client/context/context'
 import { TextDocumentIdentifier } from '../../client/types/textDocument'
-import { Context } from '../../environment/context/context'
 import { MessageConnection } from '../../jsonrpc2/connection'
 import { InitializeParams, Settings, TextDocumentDecoration } from '../../protocol'
